Created single checkbox widget
It's just a simple boolean box. The major difference between this and a checklist is that this one doesn't hold a title. Checklists are meant to be a logical group of data, so they must hold a title. Also, creating a checklist, when AbsCurses is used, creates a simple group of checkboxes with a title.