Handle missing modules
Not everyone using X Windows uses GTK (nor wants it).  To help support these
people, protect the call to 'require "abstk.AbsGtk"' to allow it to fail
without aborting the program.

Do the same with those that want to use GTK, but don't want to bother with
the Curses interface, just to be consistent.