Support Lua versions 5.1 and 5.2
Modify the code to run under Lua 5.1 and 5.2.  This is mostly just stubbing
out utf8.len() with string.len().  This is probably a gross hack (and I'll
understand it being rejected) but at the very least, it'll get someone up
and running until a better utf8.len() for Lua 5.1/5.2 can be found or
written.